- added challenge requirements
This is a difficult challenge. You pass the challenge if you can create the vertical and horizontal stacks required for the layout. Once you've done that, you can also get extra credit if you figure out how to layout the small details.

Bronze Medal: If you can figure out how to have the layout be constrained to the Safe Areas in Portrait and Landscape. (Notice the distance to the edges is different between landscape and portrait.

Silver Medal: If you can figure out how to indent the Label with the 0 so that there is 20px distance from the right edge of the screen.

Gold Medal: If you can figure out how to make the 0 button take up twice as much width as the "." and "=" buttons.

Name    :   Maricel Louise Sumulong
Date    :   July 13, 2022 PDT
E-mail  :   mllsumulong@gmail.com